Delay showing Terms & Conditions dialog
