+++ ankihub/gui/addons.py
@@ -3,12 +3,11 @@ Handles problems with add-on updates and deletions."""
 from concurrent.futures import Future
 from typing import Any, Callable
 
-import aqt
 from anki.hooks import wrap
 from aqt import addons
 from aqt.addons import DownloaderInstaller
 
-from .. import LOGGER
+from .utils import run_with_delay_when_progress_dialog_is_open
 
 
 def setup_addons():
@@ -61,22 +60,4 @@ def _with_delay_when_progress_dialog_is_open(*args, **kwargs) -> Any:
     _old: Callable = kwargs["_old"]
     del kwargs["_old"]
 
-    def wrapper():
-        LOGGER.info("Calling with_delay_when_progress_dialog_is_open._old")
-        _old(*args, **kwargs)
-
-        # the documentation of aqt.mw.progress.timer says that the timer has to be deleted to
-        # prevent memory leaks
-        timer.deleteLater()
-
-    # aqt.mw.progress.timer is there for creating "Custom timers which avoid firing while a progress dialog is active".
-    # It's better to use a large delay value because there is a 0.5 second time window in which
-    # the func can be called even if the progress dialog is not closed yet.
-    # See https://github.com/ankitects/anki/blob/d9f1e2264804481a2549b23dbc8a530857ad57fc/qt/aqt/progress.py#L261-L277
-    timer = aqt.mw.progress.timer(
-        ms=2000,
-        func=wrapper,
-        repeat=False,
-        requiresCollection=True,
-        parent=aqt.mw,
-    )
+    run_with_delay_when_progress_dialog_is_open(_old, *args, **kwargs)
